Replace `f64` suffix with float literal `.0`
Using `.0` for whole numbers is clearer than using `f64`, e.g. `100.0`
vs. `100f64`. This change also makes it easier later to replace `f64`
types with a different type via search-and-replace.

Change PAC/POD constants from `u32` to `f64`.

+++ packages/sci-cream/src/constants.rs
@@ -1,21 +1,21 @@
 pub const ABV_TO_ABW_RATIO: f64 = 0.789;
 
-pub const GLUCOSE_POD: u32 = 70;
-pub const FRUCTOSE_POD: u32 = 170;
-pub const GALACTOSE_POD: u32 = 0; // @todo
-pub const SUCROSE_POD: u32 = 100;
-pub const LACTOSE_POD: u32 = 16;
-pub const MALTOSE_POD: u32 = 80; // @todo
+pub const GLUCOSE_POD: f64 = 70.0;
+pub const FRUCTOSE_POD: f64 = 170.0;
+pub const GALACTOSE_POD: f64 = 0.0; // @todo
+pub const SUCROSE_POD: f64 = 100.0;
+pub const LACTOSE_POD: f64 = 16.0;
+pub const MALTOSE_POD: f64 = 80.0; // @todo
 
-pub const GLUCOSE_PAC: u32 = 190;
-pub const FRUCTOSE_PAC: u32 = 190;
-pub const GALACTOSE_PAC: u32 = 190; // @todo
-pub const SUCROSE_PAC: u32 = 100;
-pub const LACTOSE_PAC: u32 = 100;
-pub const MALTOSE_PAC: u32 = 100; // @todo
+pub const GLUCOSE_PAC: f64 = 190.0;
+pub const FRUCTOSE_PAC: f64 = 190.0;
+pub const GALACTOSE_PAC: f64 = 190.0; // @todo
+pub const SUCROSE_PAC: f64 = 100.0;
+pub const LACTOSE_PAC: f64 = 100.0;
+pub const MALTOSE_PAC: f64 = 100.0; // @todo
 
-pub const SALT_PAC: u32 = 585;
-pub const ALCOHOL_PAC: u32 = 740;
+pub const SALT_PAC: f64 = 585.0;
+pub const ALCOHOL_PAC: f64 = 740.0;
 
 pub const CACAO_FAT_HF: f64 = 0.9;
 pub const COCOA_SOLIDS_HF: f64 = 1.8;
